Changed the drawer + clean the code
